Added a digest-stream option to calculate hashes on pipes
MD5/SHA1/SHA256 are calculated for a given input stream. The digest is
then written to an output file. One use case for this would be writing a
file to disk and simultaneously calculating an MD5 on the data as it is
already in memory. Then you don't need to re-read the data.