Major update to BGZF processing
BGFZ blocks are now cached, not just the cOffset/uOffset requests.
Blocks are cached using their compressed file offset. This is handled
automatically.

TabixQueryIterator will now only return a single line once. In the past,
there was a chance that a line could be returned more than once if it
was part of multiple bins. Now the line offset in the file is tracked, 
so any given line can't be duplicated.