Updated VCF annotation to include more format checks (no blank values)
It used to be possible to write a VCF attribute that had an empty value
("" which ends up as KEY= ... it's the empty value after the = that is
the issue). This isn't valid per the VCF spec, so a new Exception is
thrown if this is tried.

Missing values (".") are allowed, so this is now a special case.

Flags (empty values) are also allowed (KEY, w/o the equals), so this is
also a special case.

Empty values were largely the result of other annotation commands
pulling in empty values (for example a blank value in a Tabix record).