Update to allow for vcf-ann --in-file to include comma delimited values
If a variant crosses multiple genes, each gene could be annotated in an
INFO tag. If you are then using this field to match to a gene list, you
would miss a variant that is part of two annotated "genes", even if one
of them was part of the gene list.