Overly elaborate method for removing duplicate records from events
When removing records based on events, you probably want to keep the
records that are matched between different VCF inputs (not-rescued). So,
there is now an option for how to sort/prioritize records that have the
same event.

It's way too complicated, but thorough. You can extract different info
fields and sort them. Sorting is handled in a type specific manner
(strings, integer, etc...). This is the overly complicated part. It's
necessary if you want to be generic, but complicated.