Reduced mem requirements when NOT using an EVENT filter
When using a unique EVENT value (SV VCFs), the program keeps all
variants in memory and writes them at the end. This was too much for a
normal germline VCF (many variants). So now the in-memory buffer is only
used for exporting VCFs that explicitly use the unique EVENT option.