Added a bedgraph annotator for tab-annotate
This will add mean/median/min/max values for regions in a tabix-indexed
file from a bedgraph file.

Bedgraphs are non-overlapping regions with a score

chrom, start, end, value

Because they are non-overlapping, they compress values into contiguous
regions

chr1 0 1 1
chr1 1 2 1
chr1 2 3 1

becomes:
chr1 0 3 1

This annotation source really only makes sense when using a tabix input
like a BED file, so this is not a VCF-level annotation (even though you
technically could use a bedgraph to annotate a VCF file, I'm not sure if
you would ever want to).

The use-case for this is to add mappability scores to BED regions.
Mappability is calculated based on the number of times a given read will
align to the genome. Because the bedgraph compresses the values, we
can't just use the normal tabix-BED input annotator. We have to do a
little math to make sure the scores are spread across all bases in the
region.