Added WPS calculation
WPS is the difference between spanning and overlapping reads across a
window. The purpose is to find peaks of coverage (primarily in cfDNA).

If there are more reads that span a window, then that window is more
likely to be nucleosome protected. The next step is to find the peak
values to identify the periodicity of the nucleosome protection.

xxxxxxxx      xxxxxxxx
      |-------|
     @@@@@@@@@@@

This would have a WPS score of (1-2) = -1