Added initial bam-varcall
This uses a negative binomial (Pascal) distribution to predict variant
calls.

For germline/diploid calls, the NB is used to predict if an allele is
homozygous, heterozygous, or in the background range.

For tumor calls, the NB is used to predict if an allele is above
background at the given allele frequency. (A normal sample can also be
used to subtract out het/hom germline calls)