Added vcf-sample-export to write FORMAT data in a TSV by sample.
This also caused a small change to how samples are identified in VCFs.
Before, if you used a number for a sample (eg: 1), it was assumed to
mean the first sample. However, the sample itself could be named "1",
but might be the 2nd or 3rd or... sample. Now, we will look at the
samples names first. If a sample is named "1", that is the one that is
reported back. If there is no sample named "1", then that is used as a
numeric index.