Initial import of bam-pir - Phase Informative Reads
Given a BAM file (and optional VCF), calculate which variants are
potentially informative for haplotype phasing. This will look for all
variants (or variants from a VCF) that are heterozygous. This will then
write out both the depth of the calls and the names of the reads that
contain that variant. This list can then be inverted to create a variant
list for each read. Then reads with similar genotypes can be combined to
form haplotype blocks.