format keys update -- added --dosage
added --dosage to convert GT values to dosage values (0,1,2). This
doesn't do anything other than look at the GT value and compare it to
the alt alleles.

But, in adding this, I realized that adding a fixed formatKeys for each
VCFRecord broke a number of annotations where format annotations were
getting added and not updating the formatKeys.

So, I reverted back to using the first sample's keys as the canonical
reference for keys we should display. But -- I also made it so that when
parsing a format field, we will store empty values when appropriate.
This should finally fix this issue.